requirements.txt
flask
pyspark
pandas
numpy
matplotlib
seaborn
google-generativeai
python-dotenv
werkzeug
openpyxl
